# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NutriTrack API
  version: "0.1.0"
  description: >
    Calorie & nutrition tracking API. All endpoints except /health and
    /auth/* require a Bearer access token. Every personal-data endpoint is
    scoped to the authenticated user server-side; client-supplied IDs are
    never trusted for ownership.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Food:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        servingSize: { type: string }
        servingUnit: { type: string, enum: [G, ML, PIECE, CUP, TBSP, TSP, OZ] }
        calories: { type: string }
        proteinG: { type: string }
        carbG: { type: string }
        fatG: { type: string }
        source: { type: string, enum: [SYSTEM, USER, IMPORTED, AI_ESTIMATED] }
        ownerId: { type: string, format: uuid, nullable: true }
    FoodLog:
      type: object
      properties:
        id: { type: string, format: uuid }
        foodNameSnapshot: { type: string }
        caloriesSnapshot: { type: string }
        proteinGSnapshot: { type: string }
        carbGSnapshot: { type: string }
        fatGSnapshot: { type: string }
        quantity: { type: string }
        meal: { type: string, enum: [BREAKFAST, LUNCH, DINNER, SNACK, CUSTOM] }
        loggedAt: { type: string, format: date-time }
        entrySource: { type: string, enum: [MANUAL, AI_ESTIMATED] }
...
paths:
...
  /foods:
    get:
      summary: Search your own foods + system foods
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 50 }
      responses:
        "200":
          description: Matching foods
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Food" }
    post:
      summary: Create a custom food (owned by you)
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Food" }
...
  /logs:
    get:
      summary: List your food logs for a given date (in your profile timezone)
      parameters:
        - in: query
          name: date
          required: true
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Logs for that day
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FoodLog" }
    post:
      summary: Log a food (nutrition is snapshotted at write time)
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FoodLog" }
        "404": { description: Food not found }
...
  /progress:
    get:
      summary: Daily calorie/macro rollup against the goal in effect each day
      parameters:
        - in: query
          name: from
          required: true
          schema: { type: string, format: date }
        - in: query
          name: to
          required: true
          schema: { type: string, format: date }
      responses:
        "200": { description: Daily rollup + weight history in range }
        "400": { description: Range exceeds 366 days }
  /ai/parse:
    post:
      summary: Parse free-text food description (untrusted output, not saved yet)
      responses:
        "201": { description: Parsed items for review, plus a requestId }
        "429": { description: Daily AI parse limit reached }
  /ai/parse/{id}/accept:
    post:
      summary: Commit reviewed/edited AI-parsed items as real, snapshotted logs
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201": { description: Logs created }
        "409": { description: This request was already accepted }
...